Chore - refactor Amazon Q Transform service utils
