+++ packages/core/src/amazonqGumby/chat/session/session.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { TransformationCandidateProject } from '../../../codewhisperer/service/transformByQHandler'
+import { TransformationCandidateProject } from '../../../codewhisperer/service/transformation/transformByQSharedHandler'
 
 export enum ConversationState {
     IDLE,
