+++ packages/core/src/codewhisperer/service/transformation/transformByQApiHandler.ts
@@ -2,277 +2,35 @@
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0
  */
-
 import {
-    BuildSystem,
-    JDKVersion,
+    FolderInfo,
     sessionPlanProgress,
     StepProgress,
     transformByQState,
     TransformByQStoppedError,
     ZipManifest,
-} from '../models/model'
-import * as codeWhisperer from '../client/codewhisperer'
+} from '../../models/model'
+import * as codeWhisperer from '../../client/codewhisperer'
 import * as crypto from 'crypto'
-import { getLogger } from '../../shared/logger'
-import { CreateUploadUrlResponse } from '../client/codewhispereruserclient'
-import { sleep } from '../../shared/utilities/timeoutUtils'
-import * as CodeWhispererConstants from '../models/constants'
+import { getLogger } from '../../../shared/logger'
+import { CreateUploadUrlResponse } from '../../client/codewhispereruserclient'
+import { sleep } from '../../../shared/utilities/timeoutUtils'
+import * as CodeWhispererConstants from '../../models/constants'
 import * as fs from 'fs-extra'
 import * as path from 'path'
 import * as os from 'os'
 import * as vscode from 'vscode'
-import { spawnSync } from 'child_process' // Consider using ChildProcess once we finalize all spawnSync calls
 import AdmZip from 'adm-zip'
-import globals from '../../shared/extensionGlobals'
-import {
-    CodeTransformJavaSourceVersionsAllowed,
-    CodeTransformMavenBuildCommand,
-    telemetry,
-} from '../../shared/telemetry/telemetry'
-import { codeTransformTelemetryState } from '../../amazonqGumby/telemetry/codeTransformTelemetryState'
-import {
-    calculateTotalLatency,
-    javapOutputToTelemetryValue,
-    JDKToTelemetryValue,
-} from '../../amazonqGumby/telemetry/codeTransformTelemetry'
-import { MetadataResult } from '../../shared/telemetry/telemetryClient'
-import request from '../../common/request'
-import { ToolkitError } from '../../shared/errors'
+import globals from '../../../shared/extensionGlobals'
+import { telemetry } from '../../../shared/telemetry/telemetry'
+import { codeTransformTelemetryState } from '../../../amazonqGumby/telemetry/codeTransformTelemetryState'
+import { calculateTotalLatency } from '../../../amazonqGumby/telemetry/codeTransformTelemetry'
+import { MetadataResult } from '../../../shared/telemetry/telemetryClient'
+import request from '../../../common/request'
 
-import {
-    JDK11VersionNumber,
-    JDK8VersionNumber,
-    projectSizeTooLargeMessage,
-} from '../../amazonqGumby/chat/controller/messenger/stringConstants'
-import {
-    NoJavaProjectsFoundError,
-    NoMavenJavaProjectsFoundError,
-    NoOpenProjectsError,
-    ZipExceedsSizeLimitError,
-} from '../../amazonqGumby/errors'
-
-export interface TransformationCandidateProject {
-    name: string
-    path: string
-    JDKVersion?: JDKVersion
-}
-
-export interface FolderInfo {
-    path: string
-    name: string
-}
-
-// log project details silently
-export async function validateAndLogProjectDetails() {
-    let reason,
-        result,
-        codeTransformLocalJavaVersion,
-        codeTransformPreValidationError = undefined
-    try {
-        const openProjects = await getOpenProjects()
-        const validProjects = await validateOpenProjects(openProjects, true)
-        if (validProjects.length > 0) {
-            const firstProjectJavaVersion = validProjects[0].JDKVersion!
-            codeTransformLocalJavaVersion = JDKToTelemetryValue(
-                firstProjectJavaVersion
-            ) as CodeTransformJavaSourceVersionsAllowed
-        }
-    } catch (err: any) {
-        result = MetadataResult.Fail
-        reason = err?.code
-        codeTransformPreValidationError = err?.name
-    } finally {
-        if (result || reason || codeTransformLocalJavaVersion || codeTransformPreValidationError) {
-            telemetry.codeTransform_projectDetails.emit({
-                passive: true,
-                codeTransformSessionId: codeTransformTelemetryState.getSessionId(),
-                codeTransformLocalJavaVersion,
-                codeTransformPreValidationError,
-                result: result ?? MetadataResult.Pass,
-                reason,
-            })
-        }
-    }
-}
-
-export function throwIfCancelled() {
-    if (transformByQState.isCancelled()) {
-        throw new TransformByQStoppedError()
-    }
-}
-
-export async function getOpenProjects(): Promise<TransformationCandidateProject[]> {
-    const folders = vscode.workspace.workspaceFolders
-
-    if (folders === undefined) {
-        throw new NoOpenProjectsError()
-    }
-
-    const openProjects: TransformationCandidateProject[] = []
-    for (const folder of folders) {
-        openProjects.push({
-            name: folder.name,
-            path: folder.uri.fsPath,
-        })
-    }
-
-    return openProjects
-}
-
-async function getJavaProjects(projects: TransformationCandidateProject[]) {
-    const javaProjects = []
-    for (const project of projects) {
-        const projectPath = project.path
-        const javaFiles = await vscode.workspace.findFiles(
-            new vscode.RelativePattern(projectPath!, '**/*.java'),
-            '**/node_modules/**',
-            1
-        )
-        if (javaFiles.length > 0) {
-            javaProjects.push(project)
-        }
-    }
-    return javaProjects
-}
-
-async function getMavenJavaProjects(javaProjects: TransformationCandidateProject[]) {
-    const mavenJavaProjects = []
-
-    for (const project of javaProjects) {
-        const projectPath = project.path
-        const buildSystem = await checkBuildSystem(projectPath!)
-        if (buildSystem === BuildSystem.Maven) {
-            mavenJavaProjects.push(project)
-        }
-    }
-
-    return mavenJavaProjects
-}
-
-async function getProjectsValidToTransform(
-    mavenJavaProjects: TransformationCandidateProject[],
-    onProjectFirstOpen: boolean = true
-) {
-    const projectsValidToTransform: TransformationCandidateProject[] = []
-    for (const project of mavenJavaProjects) {
-        let detectedJavaVersion = undefined
-        const projectPath = project.path
-        const compiledJavaFiles = await vscode.workspace.findFiles(
-            new vscode.RelativePattern(projectPath!, '**/*.class'),
-            '**/node_modules/**',
-            1
-        )
-        if (compiledJavaFiles.length > 0) {
-            const classFilePath = `${compiledJavaFiles[0].fsPath}`
-            const baseCommand = 'javap'
-            const args = ['-v', classFilePath]
-            const spawnResult = spawnSync(baseCommand, args, { shell: false, encoding: 'utf-8' })
-            if (spawnResult.status !== 0) {
-                let errorLog = ''
-                errorLog += spawnResult.error ? JSON.stringify(spawnResult.error) : ''
-                errorLog += `${spawnResult.stderr}\n${spawnResult.stdout}`
-                getLogger().error(`CodeTransformation: Error in running javap command = ${errorLog}`)
-                let errorReason = ''
-                if (spawnResult.stdout) {
-                    errorReason = 'JavapExecutionError'
-                } else {
-                    errorReason = 'JavapSpawnError'
-                }
-                if (spawnResult.error) {
-                    const errorCode = (spawnResult.error as any).code ?? 'UNKNOWN'
-                    errorReason += `-${errorCode}`
-                }
-                if (!onProjectFirstOpen) {
-                    telemetry.codeTransform_isDoubleClickedToTriggerInvalidProject.emit({
-                        codeTransformSessionId: codeTransformTelemetryState.getSessionId(),
-                        codeTransformPreValidationError: 'NoJavaProject',
-                        codeTransformRuntimeError: errorReason,
-                        result: MetadataResult.Fail,
-                        reason: 'CannotDetermineJavaVersion',
-                    })
-                }
-            } else {
-                const majorVersionIndex = spawnResult.stdout.indexOf('major version: ')
-                const javaVersion = spawnResult.stdout.slice(majorVersionIndex + 15, majorVersionIndex + 17).trim()
-                if (javaVersion === JDK8VersionNumber) {
-                    detectedJavaVersion = JDKVersion.JDK8
-                } else if (javaVersion === JDK11VersionNumber) {
-                    detectedJavaVersion = JDKVersion.JDK11
-                } else {
-                    detectedJavaVersion = JDKVersion.UNSUPPORTED
-                    if (!onProjectFirstOpen) {
-                        telemetry.codeTransform_isDoubleClickedToTriggerInvalidProject.emit({
-                            codeTransformSessionId: codeTransformTelemetryState.getSessionId(),
-                            codeTransformPreValidationError: 'UnsupportedJavaVersion',
-                            result: MetadataResult.Fail,
-                            reason: javapOutputToTelemetryValue(javaVersion),
-                        })
-                    }
-                }
-            }
-        }
-
-        // detectedJavaVersion will be undefined if there are no .class files or if javap errors out, otherwise it will be JDK8, JDK11, or UNSUPPORTED
-        project.JDKVersion = detectedJavaVersion
-        projectsValidToTransform.push(project)
-    }
-    return projectsValidToTransform
-}
-
-/*
- * This function filters all open projects by first searching for a .java file and then searching for a pom.xml file in all projects.
- * It also tries to detect the Java version of each project by running "javap" on a .class file of each project.
- * As long as the project contains a .java file and a pom.xml file, the project is still considered valid for transformation,
- * and we allow the user to specify the Java version.
- */
-export async function validateOpenProjects(
-    projects: TransformationCandidateProject[],
-    onProjectFirstOpen: boolean = true
-) {
-    const javaProjects = await getJavaProjects(projects)
-
-    if (javaProjects.length === 0) {
-        if (!onProjectFirstOpen) {
-            void vscode.window.showErrorMessage(CodeWhispererConstants.noSupportedJavaProjectsFoundMessage)
-            telemetry.codeTransform_isDoubleClickedToTriggerInvalidProject.emit({
-                codeTransformSessionId: codeTransformTelemetryState.getSessionId(),
-                codeTransformPreValidationError: 'NoJavaProject',
-                result: MetadataResult.Fail,
-                reason: 'CouldNotFindJavaProject',
-            })
-        }
-        throw new NoJavaProjectsFoundError()
-    }
-
-    const mavenJavaProjects = await getMavenJavaProjects(javaProjects)
-    if (mavenJavaProjects.length === 0) {
-        if (!onProjectFirstOpen) {
-            void vscode.window.showErrorMessage(
-                CodeWhispererConstants.noPomXmlFoundMessage.replace(
-                    'LINK_HERE',
-                    CodeWhispererConstants.linkToPrerequisites
-                )
-            )
-            telemetry.codeTransform_isDoubleClickedToTriggerInvalidProject.emit({
-                codeTransformSessionId: codeTransformTelemetryState.getSessionId(),
-                codeTransformPreValidationError: 'NonMavenProject',
-                result: MetadataResult.Fail,
-                reason: 'NoPomFileFound',
-            })
-        }
-        throw new NoMavenJavaProjectsFoundError()
-    }
-
-    /*
-     * These projects we know must contain a pom.xml and a .java file
-     * here we try to get the Java version of each project so that we
-     * can pre-select a default version in the QuickPick for them.
-     */
-    const projectsValidToTransform = await getProjectsValidToTransform(mavenJavaProjects, onProjectFirstOpen)
-
-    return projectsValidToTransform
-}
+import { projectSizeTooLargeMessage } from '../../../amazonqGumby/chat/controller/messenger/stringConstants'
+import { ZipExceedsSizeLimitError } from '../../../amazonqGumby/errors'
+import { throwIfCancelled, writeLogs } from './transformByQSharedHandler'
 
 export function getSha256(buffer: Buffer) {
     const hasher = crypto.createHash('sha256')
@@ -448,184 +206,6 @@ function getFilesRecursively(dir: string, isDependenciesFolder: boolean): string
         }
     })
     return files
-}
-
-// run 'install' with either 'mvnw.cmd', './mvnw', or 'mvn' (if wrapper exists, we use that, otherwise we use regular 'mvn')
-function installProjectDependencies(dependenciesFolder: FolderInfo) {
-    // baseCommand will be one of: '.\mvnw.cmd', './mvnw', 'mvn'
-    const baseCommand = transformByQState.getMavenName()
-    const modulePath = transformByQState.getProjectPath()
-
-    transformByQState.appendToErrorLog(`Running command ${baseCommand} clean install`)
-
-    // Note: IntelliJ runs 'clean' separately from 'install'. Evaluate benefits (if any) of this.
-    const args = [`-Dmaven.repo.local=${dependenciesFolder.path}`, 'clean', 'install', '-q']
-    let environment = process.env
-
-    if (transformByQState.getJavaHome() !== undefined) {
-        environment = { ...process.env, JAVA_HOME: transformByQState.getJavaHome() }
-    }
-
-    const argString = args.join(' ')
-    const spawnResult = spawnSync(baseCommand, args, {
-        cwd: modulePath,
-        shell: true,
-        encoding: 'utf-8',
-        env: environment,
-        maxBuffer: CodeWhispererConstants.maxBufferSize,
-    })
-
-    if (spawnResult.status !== 0) {
-        let errorLog = ''
-        errorLog += spawnResult.error ? JSON.stringify(spawnResult.error) : ''
-        errorLog += `${spawnResult.stderr}\n${spawnResult.stdout}`
-        transformByQState.appendToErrorLog(`${baseCommand} ${argString} failed: \n ${errorLog}`)
-        getLogger().error(
-            `CodeTransformation: Error in running Maven ${argString} command ${baseCommand} = ${errorLog}`
-        )
-        let errorReason = ''
-        if (spawnResult.stdout) {
-            errorReason = `Maven ${argString}: InstallationExecutionError`
-            /*
-             * adding this check here because these mvn commands sometimes generate a lot of output.
-             * rarely, a buffer overflow has resulted when these mvn commands are run with -X, -e flags
-             * which are not being used here (for now), but still keeping this just in case.
-             */
-            if (Buffer.byteLength(spawnResult.stdout, 'utf-8') > CodeWhispererConstants.maxBufferSize) {
-                errorReason += '-BufferOverflow'
-            }
-        } else {
-            errorReason = `Maven ${argString}: InstallationSpawnError`
-        }
-        if (spawnResult.error) {
-            const errorCode = (spawnResult.error as any).code ?? 'UNKNOWN'
-            errorReason += `-${errorCode}`
-        }
-        let mavenBuildCommand = transformByQState.getMavenName()
-        // slashes not allowed in telemetry
-        if (mavenBuildCommand === './mvnw') {
-            mavenBuildCommand = 'mvnw'
-        } else if (mavenBuildCommand === '.\\mvnw.cmd') {
-            mavenBuildCommand = 'mvnw.cmd'
-        }
-        telemetry.codeTransform_mvnBuildFailed.emit({
-            codeTransformSessionId: codeTransformTelemetryState.getSessionId(),
-            codeTransformMavenBuildCommand: mavenBuildCommand as CodeTransformMavenBuildCommand,
-            result: MetadataResult.Fail,
-            reason: errorReason,
-        })
-        throw new ToolkitError(`Maven ${argString} error`, { code: 'MavenExecutionError' })
-    } else {
-        transformByQState.appendToErrorLog(`${baseCommand} ${argString} succeeded`)
-    }
-}
-
-function copyProjectDependencies(dependenciesFolder: FolderInfo) {
-    // baseCommand will be one of: '.\mvnw.cmd', './mvnw', 'mvn'
-    const baseCommand = transformByQState.getMavenName()
-    const modulePath = transformByQState.getProjectPath()
-
-    transformByQState.appendToErrorLog(`Running command ${baseCommand} copy-dependencies`)
-
-    const args = [
-        'dependency:copy-dependencies',
-        `-DoutputDirectory=${dependenciesFolder.path}`,
-        '-Dmdep.useRepositoryLayout=true',
-        '-Dmdep.copyPom=true',
-        '-Dmdep.addParentPoms=true',
-        '-q',
-    ]
-    let environment = process.env
-    // if JAVA_HOME not found or not matching project JDK, get user input for it and set here
-    if (transformByQState.getJavaHome() !== undefined) {
-        environment = { ...process.env, JAVA_HOME: transformByQState.getJavaHome() }
-    }
-    const spawnResult = spawnSync(baseCommand, args, {
-        cwd: modulePath,
-        shell: true,
-        encoding: 'utf-8',
-        env: environment,
-        maxBuffer: CodeWhispererConstants.maxBufferSize,
-    })
-    if (spawnResult.status !== 0) {
-        let errorLog = ''
-        errorLog += spawnResult.error ? JSON.stringify(spawnResult.error) : ''
-        errorLog += `${spawnResult.stderr}\n${spawnResult.stdout}`
-        transformByQState.appendToErrorLog(`${baseCommand} copy-dependencies failed: \n ${errorLog}`)
-        getLogger().error(
-            `CodeTransformation: Error in running Maven copy-dependencies command ${baseCommand} = ${errorLog}`
-        )
-        let errorReason = ''
-        if (spawnResult.stdout) {
-            errorReason = 'Maven Copy: CopyDependenciesExecutionError'
-            if (Buffer.byteLength(spawnResult.stdout, 'utf-8') > CodeWhispererConstants.maxBufferSize) {
-                errorReason += '-BufferOverflow'
-            }
-        } else {
-            errorReason = 'Maven Copy: CopyDependenciesSpawnError'
-        }
-        if (spawnResult.error) {
-            const errorCode = (spawnResult.error as any).code ?? 'UNKNOWN'
-            errorReason += `-${errorCode}`
-        }
-        let mavenBuildCommand = transformByQState.getMavenName()
-        // slashes not allowed in telemetry
-        if (mavenBuildCommand === './mvnw') {
-            mavenBuildCommand = 'mvnw'
-        } else if (mavenBuildCommand === '.\\mvnw.cmd') {
-            mavenBuildCommand = 'mvnw.cmd'
-        }
-        telemetry.codeTransform_mvnBuildFailed.emit({
-            codeTransformSessionId: codeTransformTelemetryState.getSessionId(),
-            codeTransformMavenBuildCommand: mavenBuildCommand as CodeTransformMavenBuildCommand,
-            result: MetadataResult.Fail,
-            reason: errorReason,
-        })
-        throw new Error('Maven copy-deps error')
-    } else {
-        transformByQState.appendToErrorLog(`${baseCommand} copy-dependencies succeeded`)
-    }
-}
-
-export function getDependenciesFolderInfo(): FolderInfo {
-    const dependencyFolderName = `${CodeWhispererConstants.dependencyFolderName}${Date.now()}`
-    const dependencyFolderPath = path.join(os.tmpdir(), dependencyFolderName)
-    return {
-        name: dependencyFolderName,
-        path: dependencyFolderPath,
-    }
-}
-
-export async function writeLogs() {
-    const logFilePath = path.join(os.tmpdir(), 'build-logs.txt')
-    fs.writeFileSync(logFilePath, transformByQState.getErrorLog())
-    return logFilePath
-}
-
-export async function prepareProjectDependencies(dependenciesFolder: FolderInfo) {
-    try {
-        copyProjectDependencies(dependenciesFolder)
-    } catch (err) {
-        // continue in case of errors
-    }
-
-    try {
-        installProjectDependencies(dependenciesFolder)
-    } catch (err) {
-        void vscode.window.showErrorMessage(
-            CodeWhispererConstants.installErrorMessage.replace(
-                'LINK_HERE',
-                CodeWhispererConstants.linkToMavenTroubleshooting
-            )
-        )
-        // open build-logs.txt file to show user error logs
-        const logFilePath = await writeLogs()
-        const doc = await vscode.workspace.openTextDocument(logFilePath)
-        await vscode.window.showTextDocument(doc)
-        throw err
-    }
-
-    throwIfCancelled()
 }
 
 export async function zipCode(dependenciesFolder: FolderInfo) {
@@ -910,43 +490,4 @@ export async function pollTransformationJob(jobId: string, validStates: string[]
         }
     }
     return status
-}
-
-export async function checkBuildSystem(projectPath: string) {
-    const mavenBuildFilePath = path.join(projectPath, 'pom.xml')
-    if (fs.existsSync(mavenBuildFilePath)) {
-        return BuildSystem.Maven
-    }
-    return BuildSystem.Unknown
-}
-
-export async function getVersionData() {
-    const baseCommand = transformByQState.getMavenName() // will be one of: 'mvnw.cmd', './mvnw', 'mvn'
-    const modulePath = transformByQState.getProjectPath()
-    const args = ['-v']
-    const spawnResult = spawnSync(baseCommand, args, { cwd: modulePath, shell: true, encoding: 'utf-8' })
-
-    let localMavenVersion: string | undefined = ''
-    let localJavaVersion: string | undefined = ''
-
-    try {
-        const localMavenVersionIndex = spawnResult.stdout.indexOf('Apache Maven')
-        const localMavenVersionString = spawnResult.stdout.slice(localMavenVersionIndex + 13).trim()
-        localMavenVersion = localMavenVersionString.slice(0, localMavenVersionString.indexOf(' ')).trim()
-    } catch (e: any) {
-        localMavenVersion = undefined // if this happens here or below, user most likely has JAVA_HOME incorrectly defined
-    }
-
-    try {
-        const localJavaVersionIndex = spawnResult.stdout.indexOf('Java version: ')
-        const localJavaVersionString = spawnResult.stdout.slice(localJavaVersionIndex + 14).trim()
-        localJavaVersion = localJavaVersionString.slice(0, localJavaVersionString.indexOf(',')).trim() // will match value of JAVA_HOME
-    } catch (e: any) {
-        localJavaVersion = undefined
-    }
-
-    getLogger().info(
-        `CodeTransformation: Ran ${baseCommand} to get Maven version = ${localMavenVersion} and Java version = ${localJavaVersion} with project JDK = ${transformByQState.getSourceJDKVersion()}`
-    )
-    return [localMavenVersion, localJavaVersion]
 }
