+++ packages/core/src/codewhisperer/service/transformation/transformationHubViewProvider.ts
@@ -4,12 +4,12 @@
  */
 
 import * as vscode from 'vscode'
-import globals from '../../shared/extensionGlobals'
-import { getJobHistory, getPlanProgress } from '../commands/startTransformByQ'
-import { StepProgress, transformByQState } from '../models/model'
-import { getTransformationSteps } from './transformByQHandler'
-import { convertToTimeString } from '../../shared/utilities/textUtilities'
-import { getLogger } from '../../shared/logger'
+import globals from '../../../shared/extensionGlobals'
+import { getJobHistory, getPlanProgress } from '../../commands/startTransformByQ'
+import { StepProgress, transformByQState } from '../../models/model'
+import { getTransformationSteps } from './transformByQSharedHandler'
+import { convertToTimeString } from '../../../shared/utilities/textUtilities'
+import { getLogger } from '../../../shared/logger'
 
 export class TransformationHubViewProvider implements vscode.WebviewViewProvider {
     public static readonly viewType = 'aws.amazonq.transformationHub'
