+++ packages/core/src/test/codewhisperer/service/transformationResultsHandler.test.ts
@@ -10,7 +10,7 @@ import {
     DiffModel,
     AddedChangeNode,
     ModifiedChangeNode,
-} from '../../../codewhisperer/service/transformationResultsViewProvider'
+} from '../../../codewhisperer/service/transformation/transformationResultsViewProvider'
 import path from 'path'
 
 const getTestFilePath = (relativePathToFile: string) => {
