+++ packages/core/src/testE2E/amazonqGumby/transformByQ.test.ts
@@ -4,7 +4,11 @@
  */
 
 import assert from 'assert'
-import { getSha256, uploadArtifactToS3, zipCode } from '../../codewhisperer/service/transformByQHandler'
+import {
+    getSha256,
+    uploadArtifactToS3,
+    zipCode,
+} from '../../codewhisperer/service/transformation/transformByQApiHandler'
 import request from '../../common/request'
 import * as CodeWhispererConstants from '../../codewhisperer/models/constants'
 import * as codeWhisperer from '../../codewhisperer/client/codewhisperer'
